Add sessions package
Have added the sessions package which contains a `Login` function that
dials the gRPC server. The package also contains function that validate
the session parameters and calls the CredentialCheck method defined in
the AuthenticationServiceClient.
Have added tests to this package which should work even if the Auth
server is down. Test coverage is still only 56.52% however. Have added a
note regarding tests to the README.